0.8.5.4 Ввели дополнительные акты и шаблоны

+++ app/acts/hiddenactsbase/models.py
@@ -1,3 +1,5 @@
+from pprint import pprint
+
 from django.db import models
 from django.shortcuts import reverse
 import re
@@ -19,26 +21,11 @@ class HiddenActIS(models.Model):
 
     class Meta:
         ordering = ['act_number']
-        verbose_name = 'Акт скрытых инженерка '
+        verbose_name = 'Акт скрытых'
+        verbose_name_plural = 'Акты скрытых'
 
     def __str__(self):
         return '{} -> {}'.format(self.act_number, self.presented_work)
-
-
-class BlowDownAct(models.Model):
-    act_number = models.CharField(max_length=20, verbose_name='Ном. Акта', blank=True)
-    act_date = models.CharField(max_length=50, verbose_name='Дата Акта')
-    trassa = models.CharField(max_length=500, verbose_name='Трасса',
-                              default='от отключающего устройства Д=     мм розлив холодного водоснабжения')
-    trassa_lenght = models.CharField(max_length=50, verbose_name='Длина трассы')
-    purge_method = models.CharField(max_length=500, verbose_name='Метод продувки',
-                                    default='сетевой водой ХВС не менее 2-3 раз объема системы до "светлой воды"')
-
-    class Meta:
-        verbose_name = 'Акт промыки (продувки)'
-
-    def __str__(self):
-        return '{} -> Промывка (продувка)'.format(self.act_number)
 
 
 class ObjectActs(models.Model):
@@ -65,10 +52,33 @@ class ObjectActs(models.Model):
     designer_engineer_decree = models.CharField(max_length=200, verbose_name='Приказ проектировщика', blank=True)
     acts_instance_num = models.CharField(max_length=100, verbose_name='Кол-во экземпляров', blank=True)
     acts = models.ManyToManyField('HiddenActIS', blank=True, related_name='object_acts')
-    blow_down_act = models.ForeignKey('BlowDownAct', blank=True, null=True, on_delete=models.CASCADE)
+    # Дополнительные акты
+    # Акт промывки/продувки (w_p_act)
+    is_washing_purging_act = models.BooleanField(default=False, verbose_name='Продувка')
+    w_p_act_number = models.CharField(max_length=20, verbose_name='Номер', blank=True)
+    w_p_act_date = models.CharField(max_length=50, verbose_name='Дата', default='"15" июня 2022 г.')
+    # Акт промывки/дезинфекции (w_d_act)
+    is_washing_disinfection_act = models.BooleanField(default=False, verbose_name='Дезинфекция')
+    w_d_act_number = models.CharField(max_length=20, verbose_name='Номер', blank=True)
+    w_d_act_date = models.CharField(max_length=50, verbose_name='Дата', default='"15" июня 2022 г.')
+    w_d_disinfection_protocol_1 = models.CharField(
+        max_length=200, verbose_name='Анализ воды 1', default='А0806/64 от 15.06.2022 г.'
+    )
+    w_d_disinfection_protocol_2 = models.CharField(
+        max_length=200, verbose_name='Анализ воды 2', default='4465/2022 от 10.06.2022 г.'
+    )
+    # Акт гидравлики (h_act)
+    is_hydraulic_testing_act = models.BooleanField(default=False, verbose_name='Гидравлика')
+    h_act_number = models.CharField(max_length=20, verbose_name='Номер', blank=True)
+    h_act_date = models.CharField(max_length=50, verbose_name='Дата', default='"15" июня 2022 г.')
+    # Акт испытания канализации (s_t_act)
+    is_sewer_testing_act = models.BooleanField(default=False, verbose_name='Испыт. канализации')
+    s_t_act_number = models.CharField(max_length=20, verbose_name='Номер', blank=True)
+    s_t_act_date = models.CharField(max_length=50, verbose_name='Дата', default='"15" июня 2022 г.')
 
     class Meta:
-        verbose_name = 'Набор актов объекта '
+        verbose_name = 'Объект'
+        verbose_name_plural = 'Объекты'
 
     def __str__(self):
         return '{}'.format(", ".join([self.system_type, self.address]))
@@ -83,30 +93,32 @@ class ObjectActs(models.Model):
         return self.contractor[0:20] + '...'
 
     def get_short_contractor(self):
+        """
+        Дает короткое имя подрядчика - извлекает из кавычек
+        """
         s = re.findall(r"[\"\«\»\']\D+?[\"\«\»\']", self.contractor)
         if len(s) == 0:
             return self.contractor
         else:
             return s[0]
 
-    def update_obj(self, cleaned_obj_data, cleaned_act_data, cleaned_blow_down_act_data):
-        # главное - массив актов скрытых cleaned_act_data
-        # из-за того, что в cleaned_act_data может быть больше, равно, меньше записей, чем в объекте
-        # лучше всего будет удалить все записи и записать новые
-
+    def update_obj(self, cleaned_obj_data, cleaned_act_data):
         self.__dict__.update(cleaned_obj_data[0])
-        # удаляем
+        # заменяем
         for act in self.acts.all():
-            act.delete()
+            if len(cleaned_act_data) > 0:
+                new_act_data = cleaned_act_data.pop()
+                act.__dict__.update(new_act_data)
+                act.save()
+                print('change', act.pk)
+            else:
+                print('delete', act.pk)
+                act.delete()
         # создаем новые
-        i = 0
         for act_data in cleaned_act_data:
-            act = self.acts.create()
-            act.__dict__.update(cleaned_act_data[i])
-            i += 1
-            act.save()
-        if self.blow_down_act is not None:
-            self.blow_down_act.__dict__.update(cleaned_blow_down_act_data[0])
-            self.blow_down_act.save()
+            self.acts.create(**act_data)
         self.save()
         return self
+
+
+
